Bug Fixes, Code Cleanup
Lots of little things this time:
1.) replaced a "blind" unlink() call in `downloads.php`, replacing with
code that only deletes the given file if it exists
2.) Removed unnecessary code that creates an uploads folder - that
folder already exists, so the process of even checking for it is
unnecessary.
3.) Fixed a bug in wordcensor.php where an improper array index was
being used.
4.) Added a bit more debug reporting to the function `parse_math_tag()`.
5.) Added an "edge case" provision to the function `set_wildcards()` to
replace math operators with the proper keywords, addressing issues #202
and #172.
6.) Fixed a bug in wordcensor.php that was causing PHP errors in the
pagination function.

+++ admin/download.php
@@ -34,7 +34,8 @@ if (isset($post_vars))
     else
     {
         $fileNames = $post_vars['filenames'];
-        unlink(_DOWNLOAD_PATH_ . $zipFilename); // clear out any old zip files to prepare for the new one.
+        // clear out old zip file if it exists, to prepare for the new one.
+        if(file_exists(_DOWNLOAD_PATH_ . $zipFilename)) unlink(_DOWNLOAD_PATH_ . $zipFilename);
         $zip = new ZipArchive();
         $success = $zip->open(_DOWNLOAD_PATH_ . $zipFilename, ZipArchive::CREATE);
